% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ide-snippet.R
\name{connection_databricks_shinyapp}
\alias{connection_databricks_shinyapp}
\title{A Shiny app that can be used to construct a \code{spark_connect} statement}
\usage{
connection_databricks_shinyapp()
}
\value{
A Shiny app
}
\description{
A Shiny app that can be used to construct a \code{spark_connect} statement
}
\examples{
\dontrun{
# Launch the interactive Databricks connection configuration app
connection_databricks_shinyapp()
}
}
\keyword{internal}
