% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-sail.R
\name{dbRemoveTable,connect_sail,character-method}
\alias{dbRemoveTable,connect_sail,character-method}
\title{Remove a table from a Sail connection}
\usage{
\S4method{dbRemoveTable}{connect_sail,character}(conn, name, ..., fail_if_missing = TRUE)
}
\arguments{
\item{conn}{A Sail connection}

\item{name}{The name of the table}

\item{...}{Not used}

\item{fail_if_missing}{If \code{FALSE}, no error is raised when the table does
not exist}
}
\value{
\code{TRUE}, invisibly
}
\description{
Sail's \verb{DROP TABLE} does not drop temp views, which is what most tables are
on Sail, so this method tries \code{dropTempView()} first.
}
\keyword{internal}
