% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_databricks}
\alias{deploy_databricks}
\title{@export}
\usage{
deploy_databricks(
  appDir = NULL,
  account = NULL,
  server = NULL,
  python = NULL,
  lint = FALSE,
  version = NULL,
  cluster_id = NULL,
  host = NULL,
  token = NULL,
  ...
)
}
\description{
@export
}
