% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-install.R
\name{installed_components}
\alias{installed_components}
\title{Lists installed Python libraries}
\usage{
installed_components(list_all = FALSE)
}
\arguments{
\item{list_all}{Flag that indicates to display all of the installed packages
or only the top two, namely, \code{pyspark} and \code{databricks.connect}}
}
\value{
Returns no value, only sends information to the console. The
information includes the current versions of 'sparklyr', and 'pysparklyr',
as well as the 'Python' environment currently loaded.
}
\description{
Lists installed Python libraries
}
