% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-pyspark.R
\name{installed_versions}
\alias{installed_versions}
\title{Lists installed Python libraries}
\usage{
installed_versions(list_all = FALSE)
}
\arguments{
\item{list_all}{Flag that indicates to display all of the installed packages
or only the top two, namely, \code{pyspark} and \code{databricks.connect}}
}
\description{
Lists installed Python libraries
}
