% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-utils.R
\name{pyspark_config}
\alias{pyspark_config}
\title{Read Spark configuration}
\usage{
pyspark_config()
}
\value{
A list object with the initial configuration that will be used for
the Connect session.
}
\description{
Read Spark configuration
}
