% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-requirements.R
\name{requirements_write}
\alias{requirements_write}
\title{Write the 'requirements.txt' file}
\usage{
requirements_write(
  envname = NULL,
  destfile = "requirements.txt",
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{envname}{The name of, or path to, a Python virtual environment.}

\item{destfile}{Target path for the requirements file. Defaults to 'requirements.txt'.}

\item{overwrite}{Replace the contents of the file if it already exists?}

\item{...}{Additional arguments passed to \code{reticulate::py_list_packages()}}
}
\description{
Write the 'requirements.txt' file
}
