% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_pyspark}
\alias{install_pyspark}
\title{Installs python dependencies}
\usage{
install_pyspark(python_version = NULL, virtualenv_name = "sparklyr2")
}
\arguments{
\item{python_version}{The version of python to install if not available}

\item{virtualenv_name}{The name of the Virtual Environment to use to
install the python libraries. Defaults to "sparklyr2".}
}
\description{
Installs python dependencies
}
