% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_pyspark}
\alias{install_pyspark}
\title{Installs python dependencies}
\usage{
install_pyspark(
  envname = "r-sparklyr",
  ...,
  python_version = ">=3.9",
  new_env = identical(envname, "r-sparklyr"),
  method = c("auto", "virtualenv", "conda")
)
}
\arguments{
\item{envname}{The name of the Python Environment to use to install the
Python libraries. Defaults to "r-sparklyr".}

\item{...}{Passed on to \code{\link[reticulate:py_install]{reticulate::py_install()}}}

\item{python_version}{The version of Python to use to create the Python
environment.}

\item{new_env}{If \code{TRUE}, any existing Python virtual environment and/or
Conda environment specified by \code{envname} is deleted first.}

\item{method}{The installation method to use. If creating a new environment,
\code{"auto"} (the default) is equivalent to \code{"virtualenv"}. Otherwise \code{"auto"}
infers the installation method based on the type of Python environment
specified by \code{envname}.}
}
\description{
Installs python dependencies
}
