% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pysparklyr-package.R
\docType{package}
\name{pysparklyr-package}
\alias{pysparklyr}
\alias{pysparklyr-package}
\title{pysparklyr: Provides a 'PySpark' Back-End for the 'sparklyr' Package}
\description{
Enables 'sparklyr' to integrate with 'Spark Connect', the modern connection protocol for Apache Spark. Supports 'Spark Connect' directly as well as 'Spark Connect'-based implementations including 'Databricks Connect' and 'Snowpark Connect'. Provides a 'PySpark' wrapper to enable data manipulation, machine learning workflows, and other operations on remote Spark clusters using the 'sparklyr' interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlverse/pysparklyr}
  \item Report bugs at \url{https://github.com/mlverse/pysparklyr/issues}
}

}
\author{
\strong{Maintainer}: Edgar Ruiz \email{edgar@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
