% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pysparklyr-package.R
\docType{package}
\name{pysparklyr-package}
\alias{pysparklyr}
\alias{pysparklyr-package}
\title{pysparklyr: Provides a 'PySpark' Back-End for the 'sparklyr' Package}
\description{
It enables 'sparklyr' to integrate with 'Spark Connect', and 'Databricks Connect' by providing a wrapper over the 'PySpark' 'python' library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlverse/pysparklyr}
  \item Report bugs at \url{https://github.com/mlverse/pysparklyr/issues}
}

}
\author{
\strong{Maintainer}: Edgar Ruiz \email{edgar@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
