% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-requirements.R
\name{requirements_write}
\alias{requirements_write}
\title{Writes the 'requirements.txt' file, containing the needed Python libraries}
\usage{
requirements_write(
  envname = NULL,
  destfile = "requirements.txt",
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{envname}{The name of, or path to, a Python virtual environment.}

\item{destfile}{Target path for the requirements file. Defaults to 'requirements.txt'.}

\item{overwrite}{Replace the contents of the file if it already exists?}

\item{...}{Additional arguments passed to \code{reticulate::py_list_packages()}}
}
\value{
No value is returned to R. The output is a text file with the
list of Python libraries.
}
\description{
This is a helper function that it is meant to be used for deployments
of the document or application. By default, \code{deploy_databricks()} will run this
function the first time you use that function to deploy content to Posit Connect.
}
