% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-stop-service.R
\name{spark_connect_service_start}
\alias{spark_connect_service_start}
\alias{spark_connect_service_stop}
\title{Starts and stops Spark Connect locally}
\usage{
spark_connect_service_start(
  version = "3.4",
  scala_version = "2.12",
  include_args = TRUE,
  ...
)

spark_connect_service_stop(version = "3.4", ...)
}
\arguments{
\item{version}{Spark version to use (3.4 or above)}

\item{scala_version}{Acceptable Scala version of packages to be loaded}

\item{include_args}{Flag that indicates whether to add the additional arguments
to the command that starts the service. At this time, only the 'packages'
argument is submitted.}

\item{...}{Optional arguments; currently unused}
}
\value{
It returns messages to the console with the status of starting, and
stopping the local Spark Connect service.
}
\description{
Starts and stops Spark Connect locally
}
