% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-service.R
\name{spark_connect_service_start}
\alias{spark_connect_service_start}
\alias{spark_connect_service_stop}
\title{Starts Spark Connect locally}
\usage{
spark_connect_service_start(version = "3.4", scala_version = "2.12", ...)

spark_connect_service_stop(version = "3.4", ...)
}
\arguments{
\item{version}{Spark version to use (3.4 or above)}

\item{scala_version}{Acceptable Scala version of packages to be loaded}

\item{...}{Optional arguments; currently unused}
}
\description{
Starts Spark Connect locally
}
